Removed old version of Graph and Graph_Generator, renamed TempGraph and TempGraphGenerator to not be temp any more
